Be more conservative on updates triggered from the index
When the index changes, check whether the object of the staged file has
also changed before running another update. If the object hasn't changed
then we don't need to update.

Whilst we are at it also collect the mode bits from the index and use
these when staging hunks instead of assuming them.