Add config._refresh_staged_on_update
Defaults to true (current behaviour). Disabling will cause the staged
file to be only be updated when the index changes resulting in less jobs
and better performance. Marking as undocumented as there *may* be cases
where updates of the staged files don't happen when they should.

Also expose refresh() in case the above issue ever occurs.