feat(deprecated): only warn hard deprecations
Add deprecations to doc.

When fields are now initially deprecated, no warning messages will be
displayed. However if a deprecated field is marked with hard, then a
warning will be issued.

At the beginning of each release cycle, all deprecated fields will be
marked as hard and fields that were already marked as hard will be
removed.