feat(actions): preview inline diff on hover
By calling `Gitsigns automatic_hunk_inline` or by enabling
`Config.automatic_hunk_inline` when the cursor is on an hunk (added,
modified or removed) `preview_hunk_inline` is called automatically.
If the cursor moves inside the hunk's bounds the extmark doesn't
flicker, if it exits the hunk the extmarks are cleared automatically.
It works via an autocommand triggering on CursorMoved.
It takes `_inline2` into consideration.

`Config.automatic_hunk_inline` is an option to enable the functionality
by default on new buffers.