feat: tweak how commit buffers are processed
Previously when attaching to a commit buffer (via gitsigns or fugitive),
gitsigns would set the revision to diff against to the parent so the
signs of that commit would be displayed.

Now that we have staged signs, they are now used for that purpose, and
so the base is no longer set to the parent.