fix: derive Staged*Cul highlight correctly
The GitSignsStagedTopdeleteCul highlight (especially the foreground) was
very hard to read. It seems the issue was that
GitSignsStagedTopdeleteCul incorrectly derived from
GitSignsStagedChangeCul instead of the expected GitSignsChangedeleteCul.

So I think the fg_factor ended up being 0.5 of 0.5 equals .25,
foreground extremely dim.

Likely just a simply cut-and-paste oversight.

With this change the dim problem goes away.