docs: remove deprecated example from README
Update the README file, removing a call to the deprecated function of
gitsigns.toggle_deleted from the keymaps suggested example.

This function was deprecated at commit 8b74e560, suggesting to use
preview_hunk_inline instead. The previous commit added the relevant
function to the keymaps, but did not remove the use of the deprecated
function from there.

This commit updates the README, removing the use of the deprecated
function in the example code.