fix(blame): handle bad git-blame output
Sometimes the output from git blame isn't 100% accurate, specifically
the `<sourceline>` of the `-incremental` output.

If `blame_line()` cannot find the hunk for the blame info on a specific
line, fallback to using the nearest hunk. To indicate that this hunk
might not be correct mark it in the popup window as `guessed`.

Fixes #1332