fix(handle_blame_info): do not consider `wrap/nowrap` for `right_align/eol`
When calculating if a line should fall back to `eol` for very long lines,
window wrapping should not be considered. This is because when window is
`nowrap`, then the blame should still be `eol`.

The idea of `right_align` option is to be nicely right aligned, but fall
back to `eol` when there is not enough room. Long lines with `nowrap` will
*never* have extra room, so they should also always fall back to `eol`
mode.