fix: Close blame buffer on closure of source buffer
When the blame buffer is open and the source buffer is closed, this
results in an error

> E855: Autocommands caused command to abort

and while this closes the blame buffer, the source buffer remains open.

We can fix this by calling nvim_win_close on QuitPre. We're already
doing this on BufHidden, so can add QuitPre to the existing autocommand.